#282 move all PageObjects related tests to a subpackage
