+++ statics/src/test/java/integration/pageobjects/ElementsContainerWithManuallyInitializedFieldsTest.java
@@ -1,8 +1,9 @@
-package integration;
+package integration.pageobjects;
 
 import com.codeborne.selenide.ElementsCollection;
 import com.codeborne.selenide.ElementsContainer;
 import com.codeborne.selenide.SelenideElement;
+import integration.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.support.FindBy;
