+++ statics/src/test/java/integration/pageobjects/PageObjectWithManuallyInitializedFieldsTest.java
@@ -1,7 +1,8 @@
-package integration;
+package integration.pageobjects;
 
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
+import integration.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
