+++ statics/src/test/java/integration/pageobjects/SelenidePageFactoryTest.java
@@ -1,6 +1,7 @@
-package integration;
+package integration.pageobjects;
 
 import com.codeborne.selenide.SelenideElement;
+import integration.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.support.FindBy;
